+++ packages/app/screens/appearance/DemoCard.tsx
@@ -1,8 +1,17 @@
 import React from 'react';
 import { Card, H2, Paragraph, XStack, Button } from 'tamagui';
 
-const DemoCard = ({ isEnabled }) => (
-  <Card elevate size="$4" bordered>
+const DemoCard = () => (
+  <Card
+    elevate
+    size="$4"
+    bordered
+    width={250}
+    height={300}
+    scale={0.9}
+    hoverStyle={{ scale: 0.925 }}
+    pressStyle={{ scale: 0.875 }}
+  >
     <Card.Header padded>
       <H2>Sony A7IV</H2>
       <Paragraph theme="alt2">Now available</Paragraph>
