--- packages/app/screens/appearance/KitchenSink.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import { Box, Text } from 'native-base';
-import { DialogDemo } from '../../components/DialogDemo';
-
-const KitchenSink = () => (
-  <Box>
-    <Text>Kitchen Sink</Text>
-    <DialogDemo />
-  </Box>
-);
-
-export default KitchenSink;
