--- packages/app/screens/appearance/PaperCard.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import { Card as RNPCard } from 'react-native-paper';
-
-const PaperCard = () => (
-  <RNPCard>
-    <RNPCard.Title title="React Native Paper Card" />
-  </RNPCard>
-);
-
-export default PaperCard;
