+++ packages/app/screens/trip/TripDetails.tsx
@@ -1,29 +1,57 @@
-import React from 'react';
-import { View } from 'react-nat../../../components/details/headerder
-import { RootState } from 'store/store';
-import useTheme from '../../hook../../../components/pack_table/Tableble
-import useCustomStyles from 'app../../../store/packsStoreore';
-import DetailsComponent from './TripDetails/DetailsComponent';
-import loadStyles from './TripDetails/styles/styles';
-import { useTripDetails } from './TripDetails/hooks/useTripDetails'; // custom hook
-../../../components/details../../../components/details
-export function TripDetails() {
-  const { currentTheme ../../../themeemee();
-  const styles = useCustomStyles(loadStyles);
-  const { data, isLoading, i../../../components/ScoreContainerner(); // Use custom hook
-../../../components/weath../../../components/weather/WeatherCard
-  // console.log("🚀 ~../../../components/TripCardard~ TripDetails ~ tripId:", tripId) const { data, isLoading, error, refetch, isOwner, isError } = useFetchSingleTrip(tripId);
+import React, { useEffect } from 'react';
+import { createParam } from 'solito';
+import { useSelector, useDispatch } from 'react-redux';
+import { View } from 'react-native';
+import { RText } from '@packrat/ui';
+import { DetailsComponent } from '../../components/details';
+import { Platform, StyleSheet } from 'react-native';
+import { CLIENT_URL } from '@env';
+import useTheme from '../../hooks/useTheme';
+import useCustomStyles from 'app/hooks/useCustomStyles';
+import { useFetchSingleTrip } from 'app/hooks/singletrips';
+import useWeather from './useWeather';
+import {
+  TableContainerComponent,
+  WeatherCardComponent,
+  TripCardComponent,
+  ScoreContainerComponent,
+  loadStyles,
+} from './TripDetailsComponents';
 
-  // useEffect(() => { // if (!tripId) return../../../utils/mapFunctionsonseTrip(tripId)); // }, [dispatch, tripId]); const states = useSelector((state: RootState) => state);
-../../../hooks/useThem../../../hooks/useTheme
+const { useParam } = createParam();
+
+export function TripDetails() {
+  const { weatherObject, weatherWeek } = useWeather();
+  const { enableDarkMode, enableLightMode, isDark, isLight, currentTheme } =
+    useTheme();
+  const styles = useCustomStyles(loadStyles);
+  const [tripId] = useParam('tripId');
+
+  // console.log("🚀 ~ file: TripDetails.js:34 ~ TripDetails ~ tripId:", tripId)
+  const { data, isLoading, error, refetch, isOwner, isError } =
+    useFetchSingleTrip(tripId);
+
+  const link = `${CLIENT_URL}/trip/${tripId}`;
+
+  // useEffect(() => {
+  //   if (!tripId) return;
+  //   dispatch(fetchSingleTrip(tripId));
+  // }, [dispatch, tripId]);
+  const states = useSelector((state) => state);
+
+  const currentTrip = useSelector((state) => state.singleTrip.singleTrip);
+
   // const user = useSelector((state) => state.auth.user);
 
-  // check if user is owner of pack, and that pack and user exists // const isOwner = currentTrip && user && currentTrip.owner_id === user._id;
+  // check if user is owner of pack, and that pack and user exists
+  // const isOwner = currentTrip && user && currentTrip.owner_id === user._id;
 
-  // const isLoading = useSelector((state) => state.singleTrip.isLoading); // const error = useSelector((state) => state.singleTrip.error); // const isError = error !== null;
+  // const isLoading = useSelector((state) => state.singleTrip.isLoading);
+  // const error = useSelector((state) => state.singleTrip.error);
+  // const isError = error !== null;
 
   if (isLoading) return <RText>Loading...</RText>;
-
+  // console.log(currentTrip.osm_ref.geoJSON, 'geoJSON');
   return (
     <View
       style={[
@@ -32,12 +60,31 @@ export function TripDetails() {
       ]}
     >
       {!isError && (
-        <DetailsComponent
-          type="trip"
-          data={data}
-          isLoading={isLoading}
-          link={link}
-        />
+        <>
+          <DetailsComponent
+            type="trip"
+            data={data}
+            isLoading={isLoading}
+            error={error}
+            additionalComps={
+              <>
+                <TableContainerComponent currentPack={data?.packs} />
+                <WeatherCardComponent
+                  weatherObject={weatherObject}
+                  weatherWeek={weatherWeek}
+                  data={data}
+                />
+                <TripCardComponent
+                  data={data}
+                  weatherObject={weatherObject}
+                  currentTheme={currentTheme}
+                />
+                <ScoreContainerComponent data={data} isOwner={isOwner} />
+              </>
+            }
+            link={link}
+          />
+        </>
       )}
     </View>
   );
